Downloads/a2a_coder/mcp/gpt4/requirements.txt
openai>=1.54.0
flask==2.3.0
python-dotenv==1.0.0
requests==2.31.0